pyproject.toml
[tool.poetry]
name = "open-meteo"
version = "0.1.0"
description = ""
authors = ["ChristmasCode <alexandr.gotchenya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
loguru = "^0.7.2"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
